fix warning(s)

+++ Include/xLib/Ui/Platform/Win/MsgBox_win.inl
@@ -24,7 +24,7 @@ MsgBox::_show_impl(
     UINT        type_impl {};
     int_t       iRv       { -1 };
 
-    type_impl = internal::enums::types.toImpl( static_cast<std::ssize_t>(a_type) );
+    type_impl = static_cast<UINT>( internal::enums::types.toImpl( static_cast<std::ssize_t>(a_type) ) );
     iRv       = ::MessageBox(nullptr, a_text.c_str(), a_title.c_str(), type_impl);
     mrRv      = internal::enums::modalResults.toCross(iRv);
